Задание 1

+++ src/Bank.java
@@ -7,7 +7,7 @@ public class Bank {
         float deposit = in.nextFloat();
 //        в депозит начисляем процент из цикла, то есть переменная deposit счетчик
         System.out.print("Введите срок вклада в месяцах:");
-        float month = in.nextInt();
+        int month = in.nextInt();
         float percent = 7;
 //создаем переменную с единицей, с типом float. во-первых программа не будет ругаться на некорректный тип данных
 //если подставим единицу в формулу, то будет ошибка тк подсавили целый тип данных int
@@ -17,7 +17,7 @@ public class Bank {
         for (int i = 0; i < month; i++){
             deposit += (deposit * percent * (one / 12) / 100);
         }
-        System.out.printf("После %f месяцев сумма вклада составит: %f",month, deposit);
+        System.out.printf("После %d месяцев сумма вклада составит: %f",month, deposit);
         in.close();
